[change] added support for ADS1115 ADC
		--enhanced MQTT messages
		--enhanced DHT11 sensor reading -> TODO: find root cause
of issue
		--updated GPIO class

+++ pub_sub/ISimpleSensor.hpp
@@ -27,7 +27,7 @@ class ISimpleSensor
 
         virtual bool readDigitalValue() = 0;
 
-        virtual int readAnalogValue() = 0;
+        virtual float readAnalogValue() = 0;
 
         // Data members
